Move controller.go into internal/ctrl
